+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+ - Ability to upload images to File Storage for Quill WYSIWYG
+
 ## 3.11.1
 ### Fixed
  - Issues with the choices icons not making their way into external librarys such as Angular, etc.