+++ Changelog.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [UNRELEASED]
 ### Added
+ - Hashing and caching for `interpolate` function.
  - Ability to upload images to File Storage for Quill WYSIWYG
 
 ## 3.11.1